test skill move change

+++ src/com/example/sakurarpg/skillMove.java
@@ -5,9 +5,9 @@ public class skillMove {
 
 	float _add_angle = 0;
 	float _one_frame_speed = 0;
-	float _frame_num = 0;
+	int _frame_num = 0;
 
-	public skillMove(float add_angle, float one_frame_speed, float frame_num) {
+	public skillMove(float add_angle, float one_frame_speed, int frame_num) {
 		this._add_angle = add_angle;
 		this._one_frame_speed = one_frame_speed;
 		this._frame_num = frame_num;
